Create B
Ok now I would like a sort major change done. I would like column K to be divided by the number of fridays left in the month the day the code is ran. SO if the code was ran today it would divide the value in column K by 1 because there is only one Friday left and it would only populate one row. If it was ran last week it would divide by two and populate 2 rows since there is 2 fridays left.  This will very by the day the code is ran. So for example if the code was ran today column L would correspond to the month of august and the value of column K should then be divided by the fridays in the month of august. And that should in turn populate the amount of rows corresponding to that amount of fridays left in that month. It would then do the same for all columns each new column being the following month. The day the code is ran would correspond to column 11 and the next month would be 12 and so on. This will allow variance based on the day and even year that the code is ran.  So For august it should divide by 5 because there is 5 fridays it should also populate 5 rows with the correct date. I would Like for the date being populated in the commits sheet to be a Friday. So lets take today for example again if the code was ran today and in the column K there was the value 200 it would divide by 1 and populate 200 in column C and then it would populate 7/26/2024 but because it doesn’t meet the two week condition it would be bumped up to 8/9/2024. This is just an example but it should follow the rules I want for whenever the code Is ran. Also when u populate a date in the commits sheet it has to be a Friday date.
I would like a revised way to ensure that there is no populated rows of blanks in the commits sheet. I keep getting blank rows in the commits sheet between values from the parts sheet. Also if a value in column b is not found on the build plan sheet when it is being populated in the qty worksheet can u highlight the cell red. Also when u populate a date in the commits sheet can u make sure it’s a Friday.